fix: enums
