+++ src/assets/constants.ts
@@ -25,64 +25,6 @@ export const billingAddress = ref<Address>({
   postalCode: '',
 });
 
-export enum AppNames {
-  Shop = 'BEST SHOP',
-  AppName = 'Ecommerce-app',
-}
-
-export enum Pages {
-  Home = 'Home',
-  Catalog = 'Catalog',
-  About = 'About us',
-  Login = 'Login',
-  Registration = 'Registration',
-  Logout = 'Logout',
-}
-
-export enum Errors {
-  Common = 'This field must not contain leading or trailing whitespace',
-  Street = 'Street cannot be empty',
-  AutoLogin = 'Auto-login failed. Please log in manually.',
-  AccountExist = 'An account with this email already exists. Please log in.',
-  Registration = 'Registration failed. Please try again.',
-  DateFormat = 'The date must be in the YYYY-MM-DD format.',
-  AgeError = 'You must be at least 13 years old.',
-  NameSpace = 'Name must not contain leading or trailing whitespace',
-  NameFormat = 'The name must contain at least one character and no special characters or numbers.',
-  SurnameSpace = 'Surname must not contain leading or trailing whitespace.',
-  SurnameFormat = 'The surname must contain at least one character and no special characters or numbers.',
-  EmailSpace = 'Email address must not contain leading or trailing whitespace',
-  EmailFormat = 'Email address must contain an "@" symbol, local part and domain name.',
-  EmailSymbol = 'must contain an "@" symbol',
-  PasswordFormat = 'Password must contain at least 8 characters, uppercase and lowercase letter, number and special character',
-  PostalCodeFormat = 'The code must follow the format for the country (e.g., 12345 or A1B 2C3 for the U.S. and Canada, respectively).',
-  ProductDescription = 'Product description not available',
-  PasswordsNotMatch = 'Passwords do not match',
-}
-
-export enum SetAddress {
-  Same = 'Use the same address for billing',
-  DefaultShipping = 'Set as default shipping address',
-  DefaultBilling = 'Set as default billing address',
-}
-
-export enum Labels {
-  labelName = 'First Name',
-  labelSurname = 'Surname',
-  labelEmail = 'Email address',
-  labelPassword = 'Password',
-  labelDate = 'Date',
-  labelCode = 'Postal code',
-}
-
-export enum Placeholders {
-  placeholderName = 'First Name',
-  placeholderSurname = 'Surname',
-  placeholderEmail = 'Enter your email',
-  placeholderPassword = 'Password',
-  placeholderCode = 'Postal code',
-}
-
 export const categoriesId = [
   {
     name: 'kids',
@@ -121,13 +63,6 @@ export const availableBrands = ['New Balance', 'Reebok', 'Adidas', 'Puma', 'Nike
 export const availableSportTypes = ['football', 'fitness', 'basketball', 'casual', 'running'];
 
 export const DefaultItemsPerPage = 8;
-
-export const enum TextEditComponent {
-  saveButton = 'Save',
-  cancelButton = 'Cancel',
-  updateSuccessMessage = 'Data updated successfully',
-  updateErrorMessage = 'An error occurred while updating',
-}
 
 export const InputStar = '*';
 export const DropSign = '▼';
