+++ src/components/layout/edit-address.vue
@@ -3,7 +3,9 @@ import { defineProps, reactive } from 'vue';
 import type { EditAddressProps } from '../../types/address';
 import { usePostalCodeValidation } from '../../utils/validate-postal-code';
 import { createActionsChangeAddress, createActionsAddAddress } from '../../utils/create-actions';
-import { Labels, SetAddress, Placeholders } from '../../assets/constants';
+import { Placeholders } from '../../enums/placeholders';
+import { Labels } from '../../enums/labels';
+import { SetAddress } from '../../enums/set-address';
 import AddressForm from './address-form.vue';
 import BaseInput from './base-input.vue';
 import { useUserStore } from '../../stores/user';
