+++ src/components/layout/user-edit.vue
@@ -5,7 +5,8 @@ import { useUserStore } from '../../stores/user';
 import { useAuthStore } from '../../stores/auth';
 import { useSnackbarStore } from '../../stores/snackbar';
 import BaseInput from '../../components/layout/base-input.vue';
-import { Labels, Placeholders } from '../../assets/constants';
+import { Placeholders } from '../../enums/placeholders';
+import { Labels } from '../../enums/labels';
 import { validateName } from '../../utils/validate-name';
 import { validateSurame } from '../../utils/validate-surname';
 import { validateDate } from '../../utils/validate-date';
