+++ src/enums/app-names.ts
@@ -1,0 +1,4 @@
+export enum AppNames {
+  Shop = 'BEST SHOP',
+  AppName = 'Ecommerce-app',
+}
