+++ src/enums/labels.ts
@@ -1,0 +1,8 @@
+export enum Labels {
+  labelName = 'First Name',
+  labelSurname = 'Surname',
+  labelEmail = 'Email address',
+  labelPassword = 'Password',
+  labelDate = 'Date',
+  labelCode = 'Postal code',
+}
