+++ src/enums/pages.ts
@@ -1,0 +1,8 @@
+export enum Pages {
+  Home = 'Home',
+  Catalog = 'Catalog',
+  About = 'About us',
+  Login = 'Login',
+  Registration = 'Registration',
+  Logout = 'Logout',
+}
