+++ src/enums/placeholders.ts
@@ -1,0 +1,7 @@
+export enum Placeholders {
+  placeholderName = 'First Name',
+  placeholderSurname = 'Surname',
+  placeholderEmail = 'Enter your email',
+  placeholderPassword = 'Password',
+  placeholderCode = 'Postal code',
+}
