+++ src/enums/set-address.ts
@@ -1,0 +1,5 @@
+export enum SetAddress {
+  Same = 'Use the same address for billing',
+  DefaultShipping = 'Set as default shipping address',
+  DefaultBilling = 'Set as default billing address',
+}
