+++ src/enums/text-edit-component.ts
@@ -1,0 +1,6 @@
+export const enum TextEditComponent {
+  saveButton = 'Save',
+  cancelButton = 'Cancel',
+  updateSuccessMessage = 'Data updated successfully',
+  updateErrorMessage = 'An error occurred while updating',
+}
