+++ src/utils/validate-street.ts
@@ -1,4 +1,4 @@
-import { Errors } from '../../src/assets/constants';
+import { Errors } from '../enums/errors';
 export default function validateStreet(value: string): string {
   if (!value || !value.trim()) {
     return Errors.Street;
