+++ tests/registration-inputs.spec.ts
@@ -6,7 +6,7 @@ import RegistrationPage from '../src/pages/registrationPage/registration-page.vu
 import { createVuetify } from 'vuetify';
 import * as components from 'vuetify/components';
 import * as directives from 'vuetify/directives';
-import { Errors } from '../src/assets/constants';
+import { Errors } from '../src/enums/errors';
 
 vi.mock('vue-router', () => ({
   useRouter: vi.fn(() => ({
